+++ src/data-flow/service-registration.js
@@ -1,6 +1,7 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
 import { CommandFlowDescriptor } from './commands/command-flow-descriptor.js';
 import { ExecutionOverviewDescriptor } from './execution/execution-overview-descriptor.js';
+import { QueryFlowDescriptor } from './queries/query-flow-descriptor.js';
 
 export function addDataFlow(services) {
   if (!(services instanceof ServiceCollection)) {
@@ -9,6 +10,7 @@ export function addDataFlow(services) {
 
   services.registerSingleton('CommandFlowDescriptor', () => new CommandFlowDescriptor());
   services.registerSingleton('ExecutionOverviewDescriptor', () => new ExecutionOverviewDescriptor());
+  services.registerSingleton('QueryFlowDescriptor', () => new QueryFlowDescriptor());
 
   return services;
 }